Core/Warden: Various cleanups and warning fixes

+++ src/server/game/Warden/WardenCheckMgr.cpp
@@ -33,7 +33,7 @@ WardenCheckMgr::WardenCheckMgr()
 
 WardenCheckMgr::~WardenCheckMgr()
 {
-    for (int i = 0; i < CheckStore.size(); ++i)
+    for (uint16 i = 0; i < CheckStore.size(); ++i)
         delete CheckStore[i];
 
     for (CheckResultContainer::iterator itr = CheckResultStore.begin(); itr != CheckResultStore.end(); ++itr)
@@ -51,7 +51,7 @@ void WardenCheckMgr::LoadWardenChecks()
     }
 
     // For reload case
-    for (int i = 0; i < CheckStore.size(); ++i)
+    for (uint16 i = 0; i < CheckStore.size(); ++i)
         delete CheckStore[i];
 
     CheckStore.clear();
